requirements.txt
asgiref==3.8.1
Django==5.1.7
PyYAML==6.0.2
RandomWords==0.4.0
sqlparse==0.5.3
